requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic-settings==2.7.0
python-multipart==0.0.12
psutil==6.1.0
email-validator==2.2.0
